fix: remove duplicate constructor in TaskList and resolve compilation error
