+++ src/main/java/pranavbot/AddDeadlineCommand.java
@@ -1,3 +1,7 @@
+import pranavbot.Storage;
+import pranavbot.TaskList;
+import pranavbot.Ui;
+
 /**
  * Command that adds a Deadline task.
  */
