+++ src/main/java/pranavbot/MarkCommand.java
@@ -1,26 +1,30 @@
+import pranavbot.Command;
+import pranavbot.Storage;
+import pranavbot.TaskList;
+import pranavbot.Ui;
+
 /**
- * Command that deletes a task by index.
+ * pranavbot.Command that marks a task as done.
  */
-public class DeleteCommand extends Command {
+public class MarkCommand extends Command {
     private final String argument;
 
-    public DeleteCommand(String argument) {
+    public MarkCommand(String argument) {
         this.argument = argument.trim();
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         if (argument.isEmpty()) {
-            ui.showError("Please specify a task number to delete.");
+            ui.showError("Please specify a task number.");
             return;
         }
 
         try {
             int index = Integer.parseInt(argument) - 1;
-            Task removed = tasks.remove(index);
-            System.out.println("Noted. I've removed this task:");
-            System.out.println(removed);
-            System.out.println("Now you have " + tasks.size() + " tasks in the list.");
+            tasks.mark(index);
+            System.out.println("Nice!! I've marked this task as done:");
+            System.out.println(tasks.get(index));
             storage.save(tasks.getAll());
         } catch (NumberFormatException e) {
             ui.showError("Please enter a valid number for the task.");
